[jobless-automation] Expose target as assetSelection in GQL (#23198)
## Summary & Motivation

- Add an `asset_selection` param to `ScheduleSnap` and `Schedule` in GQL
API
- Populate `asset_selection` on both `ScheduleSnap` and `SensorSnap`
(which leads to population of corresponding fields in GQL) using the
`target` passed to `ScheduleDefinition`/`SensorDefinition` for jobless
automation.

Background: "jobless" (i.e. assets passed as `target`) schedules/sensors
create an anonymous asset job under the hood. We do not want the UI to
expose this asset job-- instead, we want it to just show the targeted
assets. There is prior art here. Before the new `target` parameter, it
was already possible to pass an `asset_selection` rather than `job` to
`SensorDefinition`, and the UI would render the target accordingly.

This PR leverages/copies this existing "jobless sensor" infrastructure
for the snap/GQL layer:

- `SensorSnap`/`GrapheneSensor` already had an `asset_selection`
parameter, so we just needed to correctly populate it with the targeted
assets when a sensor has an anonymous asset job. This allows the UI to
inspect the `assetSelection` field on a fetched sensor and render the
selected assets and hide the job for all jobless sensors.
- `ScheduleSnap`/`GrapheneSensor` did not have an `asset_selection`
parameter, but it was simple to just copy the approach already used for
sensors.

## How I Tested These Changes

New unit tests. @hellendag has also stacked UI on this that is working
in his tests:

- #23208
- #23233