Add backend support for partition wiping (#21976)
Internal companion PR: https://github.com/dagster-io/internal/pull/9993

## Summary & Motivation

The status quo is no API at all for performing partition-scoped asset
wiping. This PR (together with its companion) adds support for partition
wiping with cloud storage, but not OSS. The reason for this limitation
is that we have no way to performantly store partition-specific wipe
timestamps without the cloud-only `AssetPartitionsTable`. In the future
we may be able to get around this by storing partition ranges directly
in the DB, but there are unresolved questions around how to handle
`StaticPartitionsDefinition` in this scenario.

How this PR works:

- GraphQL
- The existing `wipeAssets` mutation has been modified (downstack in
#22249) to accept the existing `GraphenePartitionsByAssetSelector`
instead of `GrapheneAssetKey`. `GraphenePartitionsByAssetSelector`
already accepts an optional partition range.
- The `wipeAssets` resolver examines each item in the provided selector
and calls the (existing and unchanged) `DagsterInstance.wipe_assets` if
there is no attached partition range.
- `wipeAssets` calls the (new) `DagsterInstance.wipe_asset_partitions`
if there is partition info. The provided partition range is resolved
into a list of keys inside the GQL resolver--
`DagsterInstance.wipe_asset_partitions` does not operate directly on
ranges.
- Instance/storage
- `DagsterInstance.wipe_asset_partitions` (currently private) simply
forwards the call to `EventLogStorage.wipe_asset_partitions`.
- All non-cloud `EventLogStorage` immediately throw if
`wipe_asset_partitions` is called. Only `PostgresCloudEventLogStorage`
implements `wipe_asset_partitions`. This is OK because the UI for wiping
asset partitions should not appear accept for cloud UI.
- `PostgresCloudEventLogStorage.wipe_asset_partitions` deletes all rows
from `AssetMaterializationsTable`, `AssetObservationsTable`, and
`AssetPartitionsTable` matching an asset partition. This approach is
mimicking what `wipe_assets` does without the partition key qualifier.
We also set the `cached_status_data` on the `AssetKeyTable` record to
`None`. We do not touch any other fields on the `AssetKeyTable`-- this
means that the `last_materialization_time` for the asset as a whole will
remain X even if X is from a run that had all its partitions wiped
since. I was a little unsure what to do here but it would add complexity
to look up the last non-deleted partition and adjust
`last_materialization_time` back.


## How I Tested These Changes

- Internal:
- New unit test for `PostgresCloudEventLogStorage.wipe_asset_partitions`
- OSS:
- There is a test added downstack in #22249 that checks for the right
error message when wiping partitions-- in that PR the error is thrown in
the GQL resolver, here the error has been pushed down to the
`EventLogStorage`. That the test is passing indicates that we are still
returning the right error message.