[ui] Use target information for Schedules (#23208)
## Summary & Motivation

Use target information for Schedules, same as we do for Sensors. This means adding asset selection secondary queries to the schedules list and detail page.

Note that this isn't shown anywhere on the schedules list, since it would involve adding a new column and I think I need to sort that out with @braunjj.

<img width="542" alt="Screenshot 2024-07-24 at 13 29 53" src="https://github.com/user-attachments/assets/23930e32-5417-4cb0-8f3c-8d915a52f9d1">

<img width="492" alt="Screenshot 2024-07-24 at 13 29 45" src="https://github.com/user-attachments/assets/4240282c-ebdd-40f6-a4d8-17e325ca4dbf">

## How I Tested These Changes

Use sample code location provided by @smackesey that includes anonymous asset jobs on sensors and schedules. Verify that the target appears correctly in both sensor and schedule detail pages, and on the existing sensors list.