[dagster-airlift] initial commit (#23254)
Initial commit of dagster-airlift utilities that I've been using in
demos.

Added a unit test suite that goes over each of the major pieces of
functionality. Yaml parsing will come in a later PR; the cruft that I
was using for the demos should not be committed lol.

The goal here is to get us to a place where we can start iterating under
test more easily.

pyright is in a crufty state right now. Hard to reconcile what's issues
with my pyright installation vs issues with pyright pins themselves.
After a debug sesh with @smackesey , I'm going to be adding this package
to the pyright excludes for now, and once sean cleans up the current
pyright pins, I'll be putting up a PR to get this under type checking.