[1.8][external assets] accept `AssetSpec`s in `materialize` (#23216)
## Summary & Motivation

`SourceAsset`s can currently be passed to the `assets` argument of
`materialize`. This allows metadata about an upstream asset to be made
available during the execution of an asset downstream of it – e.g. to
grab the upstream `PartitionsDefinition` to resolve the upstream
partitions that the execution depends on.

This PR enables the same with `AssetSpec`s, given the impending
deprecation of `SourceAsset`.

This PR depends on `Definitions` accepting `AssetSpec`s
(https://github.com/dagster-io/dagster/pull/22961).

## How I Tested These Changes