support unserializable metadata on `AssetCheckSpec` (#23190)
## Summary & Motivation

In line with @schrockn's proposal here:
https://github.com/dagster-io/dagster/pull/22314. This makes it function
the same as `AssetSpec`.

The direct motivation for this is
https://github.com/dagster-io/dagster/pull/23191.

## How I Tested These Changes

I would have also added a test to make sure it doesn't cause any errors
on serialization, but discovered that `ExternalAssetCheck` doesn't
actually have a metadata field yet, so this metadata is not currently
serialized. I'll try to add that field as a followup if I can squeeze it
in.