[databricks] removal of legacy `databricks_api` client
Deprecated in version 0.21.0, the unofficial `databricks_api` client has
been archived with a statement to use the official Databraicks SDK for
Python. A comment was left July 24, 2023 to remove this dependency in the
next minor release, however, it appears to have been missed.

This pull request removes the legacy client authentication method, along
with the `databricks_api` dependency in `setup.py`.

To the reviewers, I would like to know if there is a way for us to gauge
the usage of this authentication method to determine downstream impact.