Synchronize sensor/schedule instigator states in LeftNav / SensorSwitch (#22954)
## Summary & Motivation

After our recent indexeddb caching changes the sensor/schedule states in
the WorkspaceContext are no longer up to date because:
1. They aren't synchronized by the apollo cache with the rest of the app
2. We no longer fetch the WorkspaceQuery on every page load, instead we
rely on the indexeddb cache and only update it if we detect that the
cache is out of date by polling code location statuses

This resulted in the LeftNav having stale statuses for
sensors/schedules.

Now that we have a resolver that can fetch instigation states
efficiently (without needing to deserialize the workspace snapshots on
the backend) we can make the LeftNav fetch updated instigation states
for the entire repository again. The Left nav is virtualized so it will
only fetch repositories within view rather than all repositories up
front.

Note that there is a big of awkwardness due to
`InstigationStateNotFoundError` being a valid type which really
indicates that we should be using the default value for the
sensor/schedule which would be the value returned from the
WorkspaceContext, so in this case we rely on it. However, when we toggle
a switch on/off elsewhere in the app, the apollo cache doesn't
automatically update the LeftNav (or any other useQueries that
previously received `InstigationStateNotFoundError`) because it never
received an instigation state. To work around this we use `useFragment`
to observe instigation states from the apollo cache directly while still
using `useQuery` to ensure we populate that cache initially.

## How I Tested These Changes
https://www.loom.com/share/8ce53ace24b847f38c23a841b6966306