[single-implicit-asset-job] allow different partitions_defs on `build_asset_job` (#23290)
## Summary & Motivation

`build_asset_job` builds a `JobDefinition` from a set of assets.
Previously, it would error if the assets targeted by the job had
different `PartitionsDefinition`s. This change adds a flag that
basically suppresses this error.

In downstream PR https://github.com/dagster-io/dagster/pull/23293, which
makes a single implicit asset job, this flag is used by the code paths
that build implicit asset jobs. Farther into the future, when we're
ready to allow users to make jobs with assets with different
`PartitionsDefinition`s, we'll remove this flag entirely and never raise
this error.


## How I Tested These Changes