[ui] Filter out runs that are too old for timeline window (#23367)
## Summary & Motivation

We are currently retrieving too many runs for the run timeline due to
our usage of `updated` time for windowing. (We should use start/end
times instead, and an upcoming fix will handle this.)

These excess completed runs show up as a big chunk at the left side of
the timeline.

To resolve this (for now), filter out runs (post-query) that don't
belong in the visible window.

## How I Tested These Changes

View timeline with lots of runs. Verify that the incorrect chunk of runs
is no longer present on the timeline.