[ui] Repair run timeline markers for timezones with non-hour offsets (#23380)
## Summary & Motivation

Resolves https://github.com/dagster-io/dagster/issues/23342.

There is currently a bug in the run timeline when rendering time
interval markers for users in timezones that do not use hour offsets
(e.g. `Asia/Kolkata`, `Australia/Adelaide`, `America/St_Johns`).

This is caused by how we calculate the initial interval marker, which is
based on a mod operation on the start time. If the user has a half-hour
offset from UTC, and we're drawing markers at every hour, we end up
drawing the time marker at the user's half-hour mark. We then label the
marker with the hour digits only -- for instance, we render a "10:30"
marker but label it as "10", thus making it appear that the timeline is
shifted incorrectly.

To fix this, use the same approach for marker generation on time markers
as we do for day markers. Create a midnight-based timestamp for the
user's timezone as a cursor, then increment by the required interval
until we reach the end of the displayed range. Then slice off any time
markers that are outside the range (e.g. they are between midnight on
that date and the `start` of the range).

## How I Tested These Changes

View the run timeline with a bunch of different timezones, including the
half-hour offset timezones named above. Verify that the date and time
markers are in the correct position, with the correct labels.