[docs] document standard asset metadata entries (#23045)
## Summary & Motivation

This documents the set of standard asset metadata entries. I marked this
an "RFC" because I'm interested in @erinkcochran87 's opinion on the
best way to fit this into this page. Right now it's a little bit
duplicative with some of the other content, and I'm not sure if it's in
the best location.

## How I Tested These Changes

---------

Co-authored-by: Erin Cochran <erin.k.cochran@gmail.com>