[ui] Replace GraphQL Explorer with GraphiQL (#23423)
## Summary & Motivation

In OSS, eliminate GraphQL Playground (which is EOL, unsupported, and
broken) and replace it with GraphiQL.

This approach uses the CDN-based static example provided in the GraphiQL
repo
([link](https://github.com/graphql/graphiql/blob/main/examples/graphiql-cdn/index.html))
but replaces the CDN host with the local template string.

The static assets are therefore checked in as part of this PR, to be
served locally, just as graphql-playground's assets were. (It seems
prudent not to depend on unpkg for this.)

<img width="914" alt="Screenshot 2024-08-05 at 12 42 12"
src="https://github.com/user-attachments/assets/3910b209-07c8-414e-95ef-4d09106d9261">


## How I Tested These Changes

Run `dagster dev -p 3333`, view GraphiQL at
`http://localhost:3333/graphql`. Verify that the UI loads and renders
correctly, and that I can successfully query the GraphQL backend.