[dagster-embedded-elt] load docstrings as default dlt resource descriptions (#23455)
## Summary

Little QoL feature for dlt assets - if a docstring is defined on the dlt
resource, we'll pull it as the default asset description.

The mapping here is a little imprecise, since technically the docstring
is describing the generic resource on the source class and not the
specific instantiation (e.g. for GitHub, what repos we're looking at),
but users can always customize that with a translator.

## Test Plan

New unit test.