[dagster-sdf] Expose SDF Table and Column Tests as Asset Checks (#23411)
## Summary & Motivation

This PR exposes some features of SDF Table and Column Tests to dagster
via AssetChecks. It is an "alpha" feature and there are many assumptions
made on the state of an sdf workspace to enable it. As such, it is
disabled by default and can be enabled via the dagster_sdf_translator
setting 'enable_asset_checks'.

In SDF, tables and columns can have tests that are authored as follows:

```
table:
  name: agg_installs_and_campaigns
  tests:
    - expect: unique_columns(["CAMPAIGN_NAME", "CAMPAIGN_NAME"])
  columns:
    - name: campaign_name
      tests:
        - expect: not_null()
        - expect: like('campaign_%')
        - expect: in_accepted_values(['campaign_a', 'campaign_b', 'campaign_c'])
```

These are expectations that columns and tables adhere to certain
conventions. By default, we use sdf's standard [testing
library](https://github.com/sdf-labs/tests).

Unlike other testing frameworks, one key innovation is that SDF tests
for all columns are [compiled into table
scans](https://docs.sdf.com/guide/data-quality/tests#tests). This allows
SDF to run multiple tests with a single scan, dramatically reducing the
cost of data tests. As a result, all tests on a table form just a single
`AssetCheck` against a table called `test_{table_name}`.

```
[Fail] Test moms_flower_shop.pub.test_agg_installs_and_campaigns
┌───────────────────────────────────────────────────────────────────────────────────────────────────────┐
│ reason                                                                                                │
╞═══════════════════════════════════════════════════════════════════════════════════════════════════════╡
│ err: column campaign_name has strings that are NOT like 'campaign_%'                                  │
│ err: column campaign_name has values outside accepted values 'campaign_a', 'campaign_b', 'campaign_c' │
│ err: columns "CAMPAIGN_NAME", "CAMPAIGN_NAME" are NOT unique                                          │
└───────────────────────────────────────────────────────────────────────────────────────────────────────┘
3 rows.
```

This naming will be updated in a later version of sdf to provide a more
explicit interface for std_lib tests.

Currently, SDF does not surface metadata in the information schema at
compile time on all available tests for a given table. As such, we
naively register an AssetCheckSpec for all tables. If no checks exist,
these will simply appear as skipped.

In the future, detailed error messages will be added to sdf's logs on
table error states, enabling descriptions of AssetChecks to benefit with
actionable error messages.

## How I Tested These Changes

See the new lineage_asset_checks workspace as well as the new
`test_asset_checks.py`