[ui] Repair box-shadow on TextInput and TextArea (#23472)
## Summary & Motivation

Repair some slightly incorrect box-shadow values on `TextInput` and
`TextArea`.

- The styled text input should have an inset box-shadow, but we appear
to have overlooked this at some point. Fixing this allows the input and
button controls to have the same default height, thus repairing the bug
that brought this to our attention.
- Note that this effectively takes the `TextInput` height down from 34px
to 32px, which may be noticeable in some parts of the app. It's more
correct and consistent, but if there are places where we're depending on
the incorrect height for alignment etc, we might notice some knock-on
effects that we should clean up.
- I also reduced the focus ring box-shadow to 2px, again to match
`Button`.
- The `TextArea` component has been ignoring its `strokeColor` prop, so
all textareas are getting the default box-shadow color. This is fixed.
- The `TextArea` now also gets a 2px focus ring.

## How I Tested These Changes

Storybook examples.