replace `GrapheneAssetNode` `isSource` with `isMaterializable` (#23401)
## Summary & Motivation

We are eliminating "source asset" as a concept in Dagster. The way we
use `isSource` in the UI is synonymous with "the asset is not
materializable". This PR removes `isSource` from `GrapheneAssetNode`,
adds `isMaterializable`, and updates the downstream frontend code.

The one user-facing change is that the asset details page will now say
"External Asset" instead of "Source Asset" when an asset is not
materializable:
![image](https://github.com/user-attachments/assets/a21cf63c-fbbc-4000-a19f-6b649cfd1f21)


## How I Tested These Changes