[sensor-upgrades][1/n] Refactor _handle_run_requests (#23242)
## Summary & Motivation

This is a simple-ish refactor which gives a single _handle_run_requests entrypoint which does both backfills and regular run requests. A small side effect is that this means we can use the threadpool submitter for submitting backfills.

The overall goal here is to make the next PR in this stack easier to implement, as we're going to be further modifying this path.

## How I Tested These Changes