[sensor-upgrades][3/n] Add automation_condition_evaluations to SensorExecutionData & co. (#23267)
## Summary & Motivation

The UserCodeAutomationConditionSensorDefinition needs to be able to return the set of evaluations that were made during its tick (regardless of if it's managed by the sensor daemon or asset daemon -- currently it simply doesn't work). This adds a new field to the result object to allow us to return that information. It is implemented as a hidden parameter of the SensorResult object so that only framework code sets it. 

While this may at first glance seem overfit to one particular type of sensor, I'm imagining a future world in which you could have regular sensors also evaluate AutomationConditions, and which would also want to emit this type of information.

This uses the same failure-handling model as the Asset Daemon, in which we write the evaluations before submitting any runs, so that if we fail to submit some runs we can always pull those evaluations back in when we resume the tick.

## How I Tested These Changes