[dagster-aws] Pipes AWS Glue Dagster run interruption handler (#23354)
## Summary & Motivation

I decided it was a good idea to add an automatic Glue job cleanup
handler as there is a change for Spark jobs to spend a lot of
unnecessary resources otherwise.

Had to rewrite the fake Glue clients to implement non-blocking job
execution via `subprocess.Popen`

Strong Inception vibes in this PR

## How I Tested These Changes
- [x] Added a test with subprocess interruption. The test runs Dagster's
`materialize` inside a `multiprocessing.Process`. Inside this process,
the fake Glue client runs the Glue job in a `subprocess.Popen`. Once the
Dagster process receives termination signal, the `PipesGlueClient`
invokes the fake glue client to terminate the `subprocess.Popen` "job".
We can register this call in the fake client and test for it.