[sensor-upgrades][4/n] Move UserCodeAutomationConditionSensorDefinition to Sensor Daemon (#23305)
## Summary & Motivation

As title -- this builds on the previous PRs which allow regular sensors to emit AutomationConditions and take advantage of checkpointing logic. Now, we flip a switch to have these user code sensors evaluated in the Sensor Daemon insteaad of the Asset Daemon.

We can consider moving the non-user code verison of these sensors to the Sensor Daemon at a later point.

## How I Tested These Changes