[ui] Hide Jobs item from nav if there are no jobs (#23521)
## Summary & Motivation

In the experimental navigation, don't show a "Jobs" top nav item if
there are no jobs in the workspace.

## How I Tested These Changes

TS, lint, jest.

Load app with default toys repo, verify that the Jobs item appears. Load
app with a repo that only contains an asset, verify that it does not
appear.