[dagster-datahub] (loosen pydantic requirement) (#23393)
## Summary & Motivation
This pydantic requirement is needlessly strict give that the usage of
Field is only in one file and uses only two parameters of it (both of
which have not changed since the Pydantic 2.0 upgrade). This needlessly
strict requirement is making it so that this won't resolve with other
packages that require pydantic 2.0. Can we please loosen this
restriction?

## How I Tested These Changes
Used the forked repo via pip

---------

Co-authored-by: Colton Padden <colton@dagsterlabs.com>