[dagster-airlift] AirflowInstance, sensor refactor (#23538)
Made the AirflowInstance vended classes more composable so we're not
throwing around raw dicts everywhere. Added some utilities to trigger
and wait for the completion of airflow runs, which are used in tests
(for now...)

Made some refactors to the sensor after realizing we don't actually want
to report in topological order... we want to report in task order. So
instead, retrieve per-task info about each execution. Better testing of
this new method is forthcoming, but existing tests already exercise the
codepath.