Update dagster uni dbt course with correct python sample file
Update 3-creating-a-partitioned-dbt-asset.md

The current code block of `dagster_university/assets/dbt.py` mentioned on this page does not reflect the state of the file during this stage of the Course. I.e. additional changes have been made which are not included in this example.

It has been updated to include these.