[dagster-airlift] mark dags as migrating (#23370)
Function that allows users to mark airflow dags as "migrating", and
injects a tag into the dag with information about the migration.
When placing the tag in the dag, there are two options:
1. Construct a new dag using essentially a shallow copy of the old dag,
and then inject this into global scope.
2. Inject a tag into the existing dag object you find in global scope
using the mutability of airflow dag's data structures.

I prefer (2) in this approach, because the surface area is way lower
than (1), and it should be relatively resistant to changes in airflow's
API other than this one tiny surface area ([which hasn't changed since
1.10, when tags were first
introduced)](https://airflow.apache.org/docs/apache-airflow/1.10.10/_modules/airflow/models/dag.html#DAG).
Unless we can figure out a reliable way to create an arbitrary copy
constructor for dags across any airflow version, we're likely to run
into brittleness with trying to reconstruct a dag from the pure object,
I think. In general, taking advantage of airflow's mutability when we
can seems like a good approach to injecting migration state.

Another point of discussion; what to do when there are no dags in scope
which the state migration object has reference to. For now I throw an
exception, but wondering if this is too harsh.

Finally, there's the question of how these tags show up in airflow's UI.
It's pretty ugly to see this json blob appear in the airflow UI after
setting the tag, but there doesn't seem to be any other data structures
we can use for this (except maybe params? But I feel more hesitant
hooking into that since it's significantly more complex implementation
wise). So might be the best we can do for now.