fix: dagster-azure AzureBlobComputeLogManager not allowing service principals (#23577)
## Summary & Motivation
The AzureBlobComputeLogManager didn't work before this. This fixes that
by allowing service principals to be used or other authentication like
managed identities. It uses the account key to generate a blob sas when
it exists, otherwise it takes the user_delegation_key.
User_delegation_keys are for service principals or managed identities.
We also set the blob permissions on the sas token as this is required,
and we fix the sas url.

A;lso during upload_blob we have to provide an overwrite argument
otherwise it doesn't work.

- closes https://github.com/dagster-io/dagster/issues/20706
- closes https://github.com/dagster-io/dagster/issues/15980

## How I Tested These Changes
Running this in production for the past 4-5 months ;)

--------

Approved to be up streamed by ASML.

---------

Co-authored-by: Colton Padden <colton@dagsterlabs.com>