[DA] Lazily compute partitionKeys for Automation tab on Asset Details page (#23559)
## Summary & Motivation

For time-partitioned assets with many historical partition keys, it is very computationally expensive to compute the string values of each of those keys. Instead, we should rely solely on partition key ranges to power the frontend.

This PR removes fields on the evaluation record which eagerly compute the number of partition keys (making it impossible to accidentally use them), and adds a separate query on the frontend to fetch the specific set of partition keys where necessary. This change resulted in a fair amount of unused code being removed.

Specific differences for the frontend:

- Now use `numCandidates` instead of `candidateSubset` (as before, if this is `None`, that represents "all partitions at the time")
- Now use the new query (`truePartitionsForAutomationConditionEvaluationNode`) to fetch the relevant partition keys instead of `trueSubset` -- sample usage can be found in the provided test

This PR breaks the frontend and requires those changes to be built on top (will coordinate with @hellendag)


## How I Tested These Changes