[dagster-airlift] create a custom operator for lakehouse example (#23597)
As discussed offline, adds a custom operator for the lakehouse example.
This should allow us to better demonstrate the step of "writing an
operator to translate dagster code."

I think this leaves us with an unclear interstitial step for this type
of asset. What's the "observe" phase look like here? Do we expect people
to just provide an assetspec? We should sketch that out in a future PR.