[ui] Code locations: new definitions page (#23651)
## Summary & Motivation

Components for the newly designed code location definitions page, which
is a stripped down browse view intended to reduce some of the list view
duplication we have throughout the app.

This is just in storybook for now, and implements only the jobs,
schedules, sensors, and resources sections. Graphs are slightly
trickier, and for ops I'm just going to transplant the existing ops
viewer. Assets are also slightly different since we'll need to
incorporate asset groups.

## How I Tested These Changes

Yarn storybook, click on each nav item. Verify searchability,
scrollability.