Add Automation Page (#23659)
This adds a "How to Automate Pipelines" guide. The idea behind this page
is that it serves as the entry point for Automation. We explain the
three ways to automate, each example offers links to a How-To guide with
examples for that concept, and potentially a concept page. Not every
idea needs a concept page. For example, Asset Sensors may only need a
Guide and Reference API documentation.

I did not include the section on Declarative Automation, but it would
also fall here.