Add "isEnabled" prop to `useAssetcatalogFiltering` to allow calling hook hook while disabling all of its side effects. (#23686)
## Summary & Motivation

In cloud we want to conditionally call this hook, or an elastic search
backed version of the hook, but we can't do that because it would break
the rules of hooks. Unfortunately the place where we want to call it
conditionally is inside of another hook used by multiple different
components and even other hooks, so going to the parent component and
making them conditionally pass 1 version of the hook or another would
not be very elegant since there would be a lot of threading of hooks
around from places which really shouldn't know anything about this hook
or its implementation. So... instead of that what we're doing is calling
both hooks, but only returning the results of the hook we're using. The
problem is that this hook has global side effects (by way of
CatalogViewProvider integration and useQueryPersistedState) so I'm
adding an `isEnabled` prop so that we disable all of these side effects
and avoid bugs.

## How I Tested These Changes

Tested with https://github.com/dagster-io/internal/pull/11040/files