[single-implicit-asset-job] add `selected_asset_keys` and `job_name` to grpc partition api params (#23527)
## Summary & Motivation

"Partition set"s are an eliminated concept that lives on like a ghost
inside our host processes and gRPC APIs. This PR is the bottom of a
stack that makes a dent at removing it.

When a job has a `PartitionsDefinition`, we also automatically generate
a `PartitionSetSnap` whose name is f"{job_name}_partition_set". This is
a relic from a past era when jobs could have multiple partition sets.
This partition set name is then referenced from host processes when they
need to get partition-related information for the job:
- The set of partitions
- The tags corresponding to a particular partition
- The config corresponding to a particular partition

This is problematic for two reasons:

- The redundant and vestigial "partition set" concept is threaded
through a bunch of our host process code
- When an asset job targets assets with different PartitionsDefinitions,
there is no partition set for it. However, we still need to be able to
get the partitions, tags, and config for subsets of assets, e.g. to
populate the Launchpad. This functionality in the Launchpad has been
broken since the single-implicit-asset-job PR was merged.

### Backwards compatibility

After this change (and some of the changes stacked on top), there will
be two kinds of code locations:
1. Code locations that don't know about these new params and never have
jobs that include assets with different `PartitionsDefinition`s.
- These code locations will continue to use the `partition_set_name`
param, which continues to be supplied by host processes, to find the
associated job and fetch the associated data.
2. Code locations that do know about these new params and may have jobs
that include assets with different `PartitionsDefinition`s.
   -  These code locations will make use of the new params.


## How I Tested These Changes

Up-stack: https://github.com/dagster-io/dagster/pull/23491