[docs-beta] Merge the new beta Docs site (#23748)
## Summary & Motivation

This adds the new beta docs site into our main repo. This is still a
work in progress and does not affect the deployment of our existing docs. This
work was originally undertaken in the #23530 branch, however, I have
brought it in as a squashed self-contained PR to keep the history clean.

## How I Tested These Changes

Built locally, BK, and Vercel.

---------

Co-authored-by: colton <cmpadden@users.noreply.github.com>