[ui] Use single hidden asset job for all single-run asset executions (#23747)
## Summary & Motivation

Related:
https://linear.app/dagster-labs/issue/FE-509/update-the-materialize-button-to-pull-jobpartitiondefinition-using-the

This PR stacks on top of #23494:

- On the backend, there is now just a single hidden asset job which
contains all assets.

- This asset job has no partition_set, but given a set of asset keys,
you can still retrieve the list of available partitions and config +
tags for a specific partition using new resolvers on the job. (Added in
@sryza's PR)

This requires some changes to expectations made on the front-end:

- We no longer check both the partition definitions AND the presence of
a partition set on the job the assets have in common. If you select
partitioned assets you will get the partitions dialog.

- When we launch a job targeting a single partition, OR open the
launchpad and allow you to choose a partition, we now use a new resolver
that is asset-specific to retrieve the config YAML + tags. (The
"partition => config YAML" resolver previously required the partition
set definition)

The last point means that there is a different config-loader query for
the hidden asset job and standard user-defined asset/ op jobs (which
still use partition set definitions). I co-located both approaches in
the same file to try to make things a bit easier to follow, but there's
a bit of code duplication.

Sidenote:

This code is super old and still supports pipelines with multiple modes.
If modes are fully gone I would love to rip that stuff out...

## How I Tested These Changes

I updated the tests and also ran through a bunch of manual tests listed
that I listed out in
https://www.notion.so/dagster/Hidden-asset-job-change-test-cases-7a96046234da46f3a0e6b2a44e349474?pvs=4.

---------

Co-authored-by: Sandy Ryza <sandy@elementl.com>
Co-authored-by: bengotow <bgotow@elementl.com>